+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 - Fix compilation error caused by `PRId32` ([#279](https://github.com/matth-x/MicroOcpp/pull/279))
 - Don't load FW-mngt. module when no handlers set ([#271](https://github.com/matth-x/MicroOcpp/pull/271))
 - Avoid creating conf when operation fails ([#290](https://github.com/matth-x/MicroOcpp/pull/290))
+- Fix whitespaces in MeterValues
 
 ## [1.0.3] - 2024-04-06